Don't check processlist when no-lock is on

+++ mydumper.c
@@ -120,7 +120,7 @@ guint binlog_connect_id= 0;
 gboolean shutdown_triggered= FALSE;
 GAsyncQueue *start_scheduled_dump;
 GMainLoop *m1;
-static GCond * ll_cond = NULL; 
+static GCond * ll_cond = NULL;
 static GMutex * ll_mutex = NULL;
 
 int errors;
@@ -422,7 +422,7 @@ void *process_queue(struct thread_data *td) {
 	} else {
 		g_message("Thread %d connected using MySQL connection ID %lu", td->thread_id, mysql_thread_id(thrconn));
 	}
-	
+
 	if(use_savepoints && mysql_query(thrconn, "SET SQL_LOG_BIN = 0")){
 		g_critical("Failed to disable binlog for the thread: %s",mysql_error(thrconn));
 		exit(EXIT_FAILURE);
@@ -471,16 +471,16 @@ void *process_queue(struct thread_data *td) {
 	    progressively waking up these threads */
 	if(less_locking){
 		g_mutex_lock(ll_mutex);
-		
+
 		while (less_locking_threads >= td->thread_id) {
 			g_cond_wait (ll_cond, ll_mutex);
 		}
-		
+
 		g_mutex_unlock(ll_mutex);
 	}
-		
+
 	for(;;) {
-		
+
 		GTimeVal tv;
 		g_get_current_time(&tv);
 		g_time_val_add(&tv,1000*1000*1);
@@ -651,7 +651,7 @@ void *process_queue_less_locking(struct thread_data *td) {
 	GString *query= g_string_sized_new(1024);
 	GString *prev_table = g_string_sized_new(100);
 	GString *prev_database = g_string_sized_new(100);
-	
+
 	for(;;) {
 		GTimeVal tv;
 		g_get_current_time(&tv);
@@ -832,24 +832,24 @@ int main(int argc, char *argv[])
 
 	time_t t;
 	time(&t);localtime_r(&t,&tval);
-	
-	//rows chunks have precedence over chunk_filesize 
+
+	//rows chunks have precedence over chunk_filesize
 	if (rows_per_file > 0 && chunk_filesize > 0){
 		chunk_filesize = 0;
 		g_warning("--chunk-filesize disabled by --rows option");
 	}
-	
+
 	//until we have an unique option on lock types we need to ensure this
 	if(no_locks || trx_consistency_only)
 		less_locking = 0;
-	
-	/* savepoints workaround to avoid metadata locking issues 
+
+	/* savepoints workaround to avoid metadata locking issues
 	   doesnt work for chuncks */
 	if(rows_per_file && use_savepoints){
 		use_savepoints = FALSE;
 		g_warning("--use-savepoints disabled by --rows");
 	}
-	
+
 	//clarify binlog coordinates with trx_consistency_only
 	if(trx_consistency_only)
 		g_warning("Using trx_consistency_only, binlog coordinates will not be accurate if you are writing to non transactional tables.");
@@ -1064,7 +1064,7 @@ void start_dump(MYSQL *conn)
 	char *p2;
 	char *p3;
 	char *u;
-	
+
 	guint64 nits[num_threads];
 	GList* nitl[num_threads];
 	int tn = 0;
@@ -1075,12 +1075,12 @@ void start_dump(MYSQL *conn)
 	guint n;
 	FILE* nufile = NULL;
 	guint have_backup_locks = 0;
-	
+
 	for(n=0;n<num_threads;n++){
 		nits[n] = 0;
 		nitl[n] = NULL;
 	}
-	
+
 	if (daemon_mode)
 		p= g_strdup_printf("%s/%d/metadata.partial", output_directory, dump_number);
 	else
@@ -1092,7 +1092,7 @@ void start_dump(MYSQL *conn)
 		g_critical("Couldn't write metadata file (%d)",errno);
 		exit(EXIT_FAILURE);
 	}
-	
+
 	if(updated_since > 0){
 		if (daemon_mode)
 			u= g_strdup_printf("%s/%d/not_updated_tables", output_directory, dump_number);
@@ -1110,45 +1110,46 @@ void start_dump(MYSQL *conn)
 	   larger than preset value, we terminate the process.
 
 	   This avoids stalling whole server with flush */
+  if(!no_locks) {
+		if (mysql_query(conn, "SHOW PROCESSLIST")) {
+			g_warning("Could not check PROCESSLIST, no long query guard enabled: %s", mysql_error(conn));
+		} else {
+			MYSQL_RES *res = mysql_store_result(conn);
+			MYSQL_ROW row;
 
-	if (mysql_query(conn, "SHOW PROCESSLIST")) {
-		g_warning("Could not check PROCESSLIST, no long query guard enabled: %s", mysql_error(conn));
-	} else {
-		MYSQL_RES *res = mysql_store_result(conn);
-		MYSQL_ROW row;
-
-		/* Just in case PROCESSLIST output column order changes */
-		MYSQL_FIELD *fields = mysql_fetch_fields(res);
-		guint i;
-		int tcol=-1, ccol=-1, icol=-1;
-		for(i=0; i<mysql_num_fields(res); i++) {
-			if (!strcasecmp(fields[i].name,"Command")) ccol=i;
-			else if (!strcasecmp(fields[i].name,"Time")) tcol=i;
-			else if (!strcasecmp(fields[i].name,"Id")) icol=i;
-		}
-		if ((tcol < 0) || (ccol < 0) || (icol < 0)) {
-			g_critical("Error obtaining information from processlist");
-			exit(EXIT_FAILURE);
-		}
-		while ((row=mysql_fetch_row(res))) {
-			if (row[ccol] && strcmp(row[ccol],"Query"))
-				continue;
-			if (row[tcol] && atoi(row[tcol])>longquery) {
-				if (killqueries) {
-					if (mysql_query(conn,p3=g_strdup_printf("KILL %lu",atol(row[icol]))))
-						g_warning("Could not KILL slow query: %s",mysql_error(conn));
-					else
-						g_warning("Killed a query that was running for %ss",row[tcol]);
-					g_free(p3);
-				} else {
-					g_critical("There are queries in PROCESSLIST running longer than %us, aborting dump,\n\t"
-						"use --long-query-guard to change the guard value, kill queries (--kill-long-queries) or use \n\tdifferent server for dump", longquery);
-					exit(EXIT_FAILURE);
+			/* Just in case PROCESSLIST output column order changes */
+			MYSQL_FIELD *fields = mysql_fetch_fields(res);
+			guint i;
+			int tcol=-1, ccol=-1, icol=-1;
+			for(i=0; i<mysql_num_fields(res); i++) {
+				if (!strcasecmp(fields[i].name,"Command")) ccol=i;
+				else if (!strcasecmp(fields[i].name,"Time")) tcol=i;
+				else if (!strcasecmp(fields[i].name,"Id")) icol=i;
+			}
+			if ((tcol < 0) || (ccol < 0) || (icol < 0)) {
+				g_critical("Error obtaining information from processlist");
+				exit(EXIT_FAILURE);
+			}
+			while ((row=mysql_fetch_row(res))) {
+				if (row[ccol] && strcmp(row[ccol],"Query"))
+					continue;
+				if (row[tcol] && atoi(row[tcol])>longquery) {
+					if (killqueries) {
+						if (mysql_query(conn,p3=g_strdup_printf("KILL %lu",atol(row[icol]))))
+							g_warning("Could not KILL slow query: %s",mysql_error(conn));
+						else
+							g_warning("Killed a query that was running for %ss",row[tcol]);
+						g_free(p3);
+					} else {
+						g_critical("There are queries in PROCESSLIST running longer than %us, aborting dump,\n\t"
+							"use --long-query-guard to change the guard value, kill queries (--kill-long-queries) or use \n\tdifferent server for dump", longquery);
+						exit(EXIT_FAILURE);
+					}
 				}
 			}
+			mysql_free_result(res);
 		}
-		mysql_free_result(res);
-	}
+  }
 
 	if (!no_locks) {
 		// Percona Backup Locks
@@ -1183,7 +1184,7 @@ void start_dump(MYSQL *conn)
 			guint retry = 0;
 			guint lock = 1;
 			int i = 0;
-			
+
 			if(db){
 				g_string_printf(query, "SELECT TABLE_SCHEMA, TABLE_NAME FROM information_schema.TABLES WHERE TABLE_SCHEMA = '%s' AND TABLE_TYPE ='BASE TABLE' AND NOT (TABLE_SCHEMA = 'mysql' AND (TABLE_NAME = 'slow_log' OR TABLE_NAME = 'general_log'))", db);
 			} else if (tables) {
@@ -1191,7 +1192,7 @@ void start_dump(MYSQL *conn)
 					dt = g_strsplit(tables[i], ".", 0);
 					dbtb = g_strdup_printf("`%s`.`%s`",dt[0],dt[1]);
 					tables_lock = g_list_append(tables_lock,dbtb);
-				}		
+				}
 			}else{
 				g_string_printf(query, "SELECT TABLE_SCHEMA, TABLE_NAME FROM information_schema.TABLES WHERE TABLE_TYPE ='BASE TABLE' AND TABLE_SCHEMA NOT IN ('information_schema', 'performance_schema', 'data_dictionary') AND NOT (TABLE_SCHEMA = 'mysql' AND (TABLE_NAME = 'slow_log' OR TABLE_NAME = 'general_log'))");
 			}
@@ -1203,7 +1204,7 @@ void start_dump(MYSQL *conn)
 				}else{
 					MYSQL_RES *res = mysql_store_result(conn);
 					MYSQL_ROW row;
-				
+
 					while ((row=mysql_fetch_row(res))) {
 						lock = 1;
 						if (tables) {
@@ -1216,8 +1217,8 @@ void start_dump(MYSQL *conn)
 						}
 						if (lock && regexstring && !check_regex(row[0],row[1]))
 							continue;
-					
-						if(lock) {					
+
+						if(lock) {
 							dbtb = g_strdup_printf("`%s`.`%s`",row[0],row[1]);
 							tables_lock = g_list_append(tables_lock,dbtb);
 						}
@@ -1240,7 +1241,7 @@ void start_dump(MYSQL *conn)
 				if(mysql_query(conn,query->str)){
 					gchar *failed_table = NULL;
 					gchar **tmp_fail;
-					
+
 					tmp_fail = g_strsplit(mysql_error(conn), "'",0);
 					tmp_fail =  g_strsplit(tmp_fail[1], ".", 0);
 					failed_table = g_strdup_printf("`%s`.`%s`", tmp_fail[0], tmp_fail[1]);
@@ -1263,7 +1264,7 @@ void start_dump(MYSQL *conn)
 			}
 			g_free(query->str);
 			g_list_free(tables_lock);
-		}else{ 
+		}else{
 			if(mysql_query(conn, "FLUSH TABLES WITH READ LOCK")) {
 				g_critical("Couldn't acquire global lock, snapshots will not be consistent: %s",mysql_error(conn));
 				errors++;
@@ -1276,7 +1277,7 @@ void start_dump(MYSQL *conn)
 		mysql_query(conn, "CREATE TABLE IF NOT EXISTS mysql.mydumperdummy (a INT) ENGINE=INNODB");
 		need_dummy_read=1;
 	}
-	
+
 	//tokudb do not support consistent snapshot
 	mysql_query(conn,"SELECT @@tokudb_version");
 	MYSQL_RES *rest = mysql_store_result(conn);
@@ -1286,7 +1287,7 @@ void start_dump(MYSQL *conn)
 		mysql_query(conn, "CREATE TABLE IF NOT EXISTS mysql.tokudbdummy (a INT) ENGINE=TokuDB");
 		need_dummy_toku_read=1;
 	}
-	
+
 	mysql_query(conn, "START TRANSACTION /*!40108 WITH CONSISTENT SNAPSHOT */");
 	if (need_dummy_read) {
 		mysql_query(conn,"SELECT /*!40001 SQL_NO_CACHE */ * FROM mysql.mydumperdummy");
@@ -1314,10 +1315,10 @@ void start_dump(MYSQL *conn)
 
 		write_snapshot_info(conn, mdfile);
 	}
-	
+
 	GThread **threads = g_new(GThread*,num_threads*(less_locking+1));
 	struct thread_data *td= g_new(struct thread_data, num_threads*(less_locking+1));
-	
+
 	if(less_locking){
 		conf.queue_less_locking = g_async_queue_new();
 		conf.ready_less_locking = g_async_queue_new();
@@ -1334,16 +1335,16 @@ void start_dump(MYSQL *conn)
 	conf.queue = g_async_queue_new();
 	conf.ready = g_async_queue_new();
 	conf.unlock_tables= g_async_queue_new();
-	
+
 	for (n=0; n<num_threads; n++) {
 		td[n].conf= &conf;
 		td[n].thread_id= n+1;
 		threads[n] = g_thread_create((GThreadFunc)process_queue,&td[n],TRUE,NULL);
 		g_async_queue_pop(conf.ready);
 	}
-	
+
 	g_async_queue_unref(conf.ready);
-	
+
 	if (trx_consistency_only){
 		g_message("Transactions started, unlocking tables");
 		mysql_query(conn, "UNLOCK TABLES /* trx-only */");
@@ -1377,11 +1378,11 @@ void start_dump(MYSQL *conn)
 		mysql_free_result(databases);
 
 	}
-	
+
 	if (!non_innodb_table){
 		g_async_queue_push(conf.unlock_tables, GINT_TO_POINTER(1));
 	}
-	
+
 	if (less_locking) {
 
 		for (non_innodb_table= g_list_first(non_innodb_table); non_innodb_table; non_innodb_table= g_list_next(non_innodb_table)) {
@@ -1397,7 +1398,7 @@ void start_dump(MYSQL *conn)
 			nitl[tn]= g_list_append(nitl[tn], dbt);
 			nits[tn] += dbt->datalength;
 		}
-		
+
 		for (n=0; n<num_threads; n++) {
 			if(nits[n] > 0){
 				g_atomic_int_inc(&non_innodb_table_counter);
@@ -1405,12 +1406,12 @@ void start_dump(MYSQL *conn)
 			}
 		}
 		g_list_free(g_list_first(non_innodb_table));
-		
+
 		if(g_atomic_int_get(&non_innodb_table_counter))
 			g_atomic_int_inc(&non_innodb_done);
 		else
 			g_async_queue_push(conf.unlock_tables, GINT_TO_POINTER(1));
-		
+
 		for (n=0; n<num_threads; n++) {
 			struct job *j = g_new0(struct job,1);
 			j->type = JOB_SHUTDOWN;
@@ -1425,7 +1426,7 @@ void start_dump(MYSQL *conn)
 		g_list_free(g_list_first(non_innodb_table));
 		g_atomic_int_inc(&non_innodb_done);
 	}
-	
+
 	for (innodb_tables= g_list_first(innodb_tables); innodb_tables; innodb_tables= g_list_next(innodb_tables)) {
 		dbt= (struct db_table*) innodb_tables->data;
 		dump_table(conn, dbt->database, dbt->table, &conf, TRUE);
@@ -1470,16 +1471,16 @@ void start_dump(MYSQL *conn)
 		get_binlogs(conn, &conf);
 	}
 	#endif
-	// close main connection 
+	// close main connection
 	mysql_close(conn);
-	
+
 	if(less_locking){
 		for (n=num_threads; n<num_threads*2; n++) {
 			g_thread_join(threads[n]);
 		}
 		g_async_queue_unref(conf.queue_less_locking);
 	}
-	
+
 	for (n=0; n<num_threads; n++) {
 		struct job *j = g_new0(struct job,1);
 		j->type = JOB_SHUTDOWN;
@@ -1573,11 +1574,11 @@ void dump_create_database(MYSQL *conn, char *database){
 void get_not_updated(MYSQL *conn){
 	MYSQL_RES *res=NULL;
 	MYSQL_ROW row;
-	
+
 	gchar *query = g_strdup_printf("SELECT CONCAT(TABLE_SCHEMA,'.',TABLE_NAME) FROM information_schema.TABLES WHERE UPDATE_TIME < NOW() - INTERVAL %d DAY",updated_since);
 	mysql_query(conn,query);
 	g_free(query);
-	
+
 	res = mysql_store_result(conn);
 	while((row = mysql_fetch_row(res)))
 		no_updated_tables = g_list_append(no_updated_tables, row[0]);
@@ -1788,7 +1789,7 @@ void dump_database(MYSQL * conn, char *database, FILE *file, struct configuratio
 	char *query;
 	mysql_select_db(conn,database);
 	if (detected_server == SERVER_TYPE_MYSQL)
-		query= g_strdup("SHOW TABLE STATUS");
+		query= g_strdup("SELECT TABLE_NAME, ENGINE, TABLE_TYPE as COMMENT FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA='%s'", database);
 	else
 		query= g_strdup_printf("SELECT TABLE_NAME, ENGINE, TABLE_TYPE as COMMENT FROM DATA_DICTIONARY.TABLES WHERE TABLE_SCHEMA='%s'", database);
 
@@ -1862,7 +1863,7 @@ void dump_database(MYSQL * conn, char *database, FILE *file, struct configuratio
 		}
 		if (!dump)
 			continue;
-		
+
 		/* Special tables */
 		if(g_ascii_strcasecmp(database, "mysql") == 0 && (g_ascii_strcasecmp(row[0], "general_log") == 0 ||
 															g_ascii_strcasecmp(row[0], "slow_log") == 0 ||
@@ -1889,7 +1890,7 @@ void dump_database(MYSQL * conn, char *database, FILE *file, struct configuratio
 		}
 		if (!dump)
 			continue;
-		
+
 		/* Green light! */
 		struct db_table *dbt = g_new(struct db_table, 1);
 		dbt->database= g_strdup(database);
@@ -2003,24 +2004,24 @@ void dump_database(MYSQL * conn, char *database, FILE *file, struct configuratio
 }
 
 void get_tables(MYSQL * conn, struct configuration *conf) {
-	
+
 	gchar **dt= NULL;
 	char *query=NULL;
 	guint i,x;
-	
+
 	for (x = 0; tables[x] != NULL; x++){
 		dt = g_strsplit(tables[x], ".", 0);
 		query= g_strdup_printf("SHOW TABLE STATUS FROM %s LIKE '%s'", dt[0], dt[1]);
-		
+
 		if (mysql_query(conn, (query))) {
 			g_critical("Error: DB: %s - Could not execute query: %s", dt[0], mysql_error(conn));
 			errors++;
 			return;
 		}
-		
+
 		MYSQL_RES *result = mysql_store_result(conn);
 		MYSQL_FIELD *fields= mysql_fetch_fields(result);
-		guint ecol= -1; 
+		guint ecol= -1;
 		guint ccol= -1;
 		for (i=0; i<mysql_num_fields(result); i++) {
 			if (!strcasecmp(fields[i].name, "Engine")) ecol= i;
@@ -2032,15 +2033,15 @@ void get_tables(MYSQL * conn, struct configuration *conf) {
 			errors++;
 			return;
 		}
-		
+
 		MYSQL_ROW row;
 		while ((row = mysql_fetch_row(result))) {
-			
+
 			int is_view=0;
 
 			if ((detected_server == SERVER_TYPE_MYSQL) && ( row[ccol] == NULL || !strcmp(row[ccol],"VIEW") ))
 				is_view=1;
-			
+
 			/* Green light! */
 			struct db_table *dbt = g_new(struct db_table, 1);
 			dbt->database= g_strdup(dt[0]);
@@ -2067,7 +2068,7 @@ void get_tables(MYSQL * conn, struct configuration *conf) {
 					view_schemas= g_list_append(view_schemas, dbt);
 				}
 			}
-		}	
+		}
 	}
 	g_free(query);
 }
@@ -2519,7 +2520,7 @@ void dump_table_data_file(MYSQL *conn, char *database, char *table, char *where,
 		return;
 	}
 	guint64 rows_count = dump_table_data(conn, (FILE *)outfile, database, table, where, filename);
-	
+
 	if (!rows_count)
 		g_message("Empty table %s.%s", database,table);
 }
@@ -2708,24 +2709,24 @@ guint64 dump_table_data(MYSQL * conn, FILE *file, char *database, char *table, c
 	char *query = NULL;
 	gchar *fcfile = NULL;
 	gchar* filename_prefix = NULL;
-	
+
 	fcfile = g_strdup (filename);
-	
+
 	if(chunk_filesize){
 		gchar** split_filename= g_strsplit(filename, ".00001.sql", 0);
 		filename_prefix= split_filename[0];
 		g_free(split_filename);
 	}
 
-	
+
 	/* Ghm, not sure if this should be statement_size - but default isn't too big for now */
 	GString* statement = g_string_sized_new(statement_size);
 	GString* statement_row = g_string_sized_new(0);
-	
+
 	/* Poor man's database code */
  	query = g_strdup_printf("SELECT %s * FROM `%s`.`%s` %s %s", (detected_server == SERVER_TYPE_MYSQL) ? "/*!40001 SQL_NO_CACHE */" : "", database, table, where?"WHERE":"",where?where:"");
 	if (mysql_query(conn, query) || !(result=mysql_use_result(conn))) {
-		//ERROR 1146 
+		//ERROR 1146
 		if(success_on_1146 && mysql_errno(conn) == 1146){
 			g_warning("Error dumping table (%s.%s) data: %s ",database, table, mysql_error(conn));
 		}else{
@@ -2782,13 +2783,13 @@ guint64 dump_table_data(MYSQL * conn, FILE *file, char *database, char *table, c
 			}
 			num_rows_st = 0;
 		}
-		
+
 		if (statement_row->len) {
 			g_string_append(statement, statement_row->str);
 			g_string_set_size(statement_row,0);
 			num_rows_st++;
 		}
-		
+
 		g_string_append(statement_row, "\n(");
 
 		for (i = 0; i < num_fields; i++) {
@@ -2851,7 +2852,7 @@ guint64 dump_table_data(MYSQL * conn, FILE *file, char *database, char *table, c
 		g_critical("Could not read data from %s.%s: %s", database, table, mysql_error(conn));
 		errors++;
 	}
-	
+
 	if (statement_row->len > 0) {
 		/* this last row has not been written out */
 		if (statement->len > 0) {
@@ -2893,13 +2894,13 @@ cleanup:
 	if (result) {
 		mysql_free_result(result);
 	}
-	
+
 	if (!compress_output){
 		fclose((FILE *)file);
 	} else {
 		gzclose((gzFile)file);
 	}
-	
+
 	if (!st_in_file && !build_empty_files) {
 		// dropping the useless file
 		if (remove(fcfile)) {
@@ -2909,10 +2910,10 @@ cleanup:
 		fcfile = g_strdup_printf("%s.sql%s", filename_prefix,(compress_output?".gz":""));
 		g_rename(filename, fcfile);
 	}
-	
+
 	g_free(filename_prefix);
 	g_free(fcfile);
-	
+
 	return num_rows;
 }
